UI Scene

+++ Mario/Mario/Environment/Ground.cpp
@@ -10,9 +10,9 @@ void Ground::SetBoxColliders(nlohmann::json groundData)
 	{
 		nlohmann::json ground = groundData.at(i);
 
-		zenObjects.push_back(new ZenObject(ground["id"], "Ground", sf::Vector2f(ground["width"]*TILE_SCALE, ground["height"] * TILE_SCALE)));
+		zenObjects.push_back(new ZenObject(ground["id"], "Ground", sf::Vector2f((float)ground["width"]*TILE_SCALE, (float)ground["height"] * TILE_SCALE)));
 
-		boxColliders.push_back(new BoxCollider2D(Vector2(ground["x"] * TILE_SCALE, ground["y"] * TILE_SCALE), Vector2(ground["width"] * TILE_SCALE, ground["height"] * TILE_SCALE), zenObjects[i], Collider::ColliderTags::GROUND));
+		boxColliders.push_back(new BoxCollider2D(Vector2((float)ground["x"] * TILE_SCALE, (float)ground["y"] * TILE_SCALE), Vector2((float)ground["width"] * TILE_SCALE, (float)ground["height"] * TILE_SCALE), zenObjects[i], Collider::ColliderTags::GROUND));
 	}
 
 	for (int i = 0; i < boxColliders.size(); i++)
